renamed files to snake_case style
