+++ main.py
@@ -1,7 +1,7 @@
-from src.School import School
-from src.Student import Student
-from src.Teacher import Teacher
-from src.Course import Course
+from src.school import School
+from src.student import Student
+from src.teacher import Teacher
+from src.course import Course
 
 # prepare data
 school = School("СЗШ 32")
