--- src/utils/Errors.py
@@ -1,2 +1,0 @@
-class CourserMaxStudentsLimitException(Exception):
-    pass
